+++ app/src/main/java/woowacourse/movie/ui/base/BaseContract.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.ui
+package woowacourse.movie.ui.base
 
 interface BaseContract {
     interface View
