+++ app/src/main/java/woowacourse/movie/ui/main/cinemaBottomSheet/CinemaListBottomSheet.kt
@@ -6,9 +6,9 @@ import com.example.domain.repositoryImpl.CinemaSampleRepository
 import woowacourse.movie.databinding.FragmentCinemaBottomSheetBinding
 import woowacourse.movie.model.CinemaState
 import woowacourse.movie.model.MovieState
-import woowacourse.movie.ui.BaseBottomSheetDialogFragment
 import woowacourse.movie.ui.adapter.CinemaListAdapter
 import woowacourse.movie.ui.adapter.itemModel.CinemaItemModel
+import woowacourse.movie.ui.base.BaseBottomSheetDialogFragment
 import woowacourse.movie.ui.reservation.MovieDetailActivity
 import woowacourse.movie.util.getParcelableCompat
