+++ app/src/main/java/woowacourse/movie/ui/main/cinemaBottomSheet/CinemaListContract.kt
@@ -2,7 +2,7 @@ package woowacourse.movie.ui.main.cinemaBottomSheet
 
 import woowacourse.movie.model.CinemaState
 import woowacourse.movie.model.MovieState
-import woowacourse.movie.ui.BaseContract
+import woowacourse.movie.ui.base.BaseContract
 
 interface CinemaListContract {
     interface View {
