+++ app/src/main/java/woowacourse/movie/ui/main/movieList/MovieListFragment.kt
@@ -6,11 +6,11 @@ import android.view.View
 import woowacourse.movie.databinding.FragmentMovieListBinding
 import woowacourse.movie.model.AdvState
 import woowacourse.movie.model.MovieState
-import woowacourse.movie.ui.BaseFragment
 import woowacourse.movie.ui.adapter.MovieListAdapter
 import woowacourse.movie.ui.adapter.itemModel.AdvItemModel
 import woowacourse.movie.ui.adapter.itemModel.MovieItemModel
 import woowacourse.movie.ui.adv.AdvDetailActivity
+import woowacourse.movie.ui.base.BaseFragment
 import woowacourse.movie.ui.main.cinemaBottomSheet.CinemaListBottomSheet
 
 class MovieListFragment : BaseFragment(), MovieListContract.View {
