+++ app/src/main/java/woowacourse/movie/ui/main/reservationList/ReservationListContract.kt
@@ -1,7 +1,7 @@
 package woowacourse.movie.ui.main.reservationList
 
 import woowacourse.movie.model.TicketsState
-import woowacourse.movie.ui.BaseContract
+import woowacourse.movie.ui.base.BaseContract
 
 interface ReservationListContract {
     interface View {
